+++ packages/web/src/pages/sign-up-page/pages/CreateEmailPage.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useState } from 'react'
+import { useCallback } from 'react'
 
 import {
   emailSchema,
@@ -48,6 +48,7 @@ import {
 import { SignUpWithMetaMaskButton } from '../components/SignUpWithMetaMaskButton'
 import { SocialMediaLoading } from '../components/SocialMediaLoading'
 import { Heading, Page } from '../components/layout'
+import { useSocialMediaLoader } from '../hooks/useSocialMediaLoader'
 
 const EmailSchema = toFormikValidationSchema(emailSchema(audiusQueryContext))
 
@@ -60,29 +61,20 @@ export const CreateEmailPage = () => {
   const dispatch = useDispatch()
   const navigate = useNavigateToPage()
   const existingEmailValue = useSelector(getEmailField)
-  const [isWaitingForSocialLogin, setIsWaitingForSocialLogin] = useState(false)
   const alreadyLinkedSocial = useSelector(getLinkedSocialOnFirstPage)
 
   const initialValues = {
     email: existingEmailValue.value ?? ''
   }
 
-  useEffect(() => {
-    // If the user goes back to this page in the middle of the flow after they linked
-    // their social on this page previously, clear the sign on state.
-    if (alreadyLinkedSocial) {
-      dispatch(resetSignOn())
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [dispatch])
-
-  const handleStartSocialMediaLogin = useCallback(() => {
-    setIsWaitingForSocialLogin(true)
-  }, [])
-
-  const handleErrorSocialMediaLogin = useCallback(() => {
-    setIsWaitingForSocialLogin(false)
-  }, [])
+  const {
+    isWaitingForSocialLogin,
+    handleStartSocialMediaLogin,
+    handleErrorSocialMediaLogin
+  } = useSocialMediaLoader({
+    resetAction: resetSignOn,
+    linkedSocialOnThisPagePreviously: alreadyLinkedSocial
+  })
 
   const handleCompleteSocialMediaLogin = useCallback(
     (result: { requiresReview: boolean; handle: string }) => {
